# api/openapi3_v1.yaml
---
openapi: 3.0.0
info:
  title: User Account Balance API
  description: |
    User Balance API for managing user account balance.
  version: 1.0.0
servers:
- url: http://localhost:8080/v1
tags:
- name: account
  description: Operations with user account
- name: reserve
  description: Operations with user reserve account
- name: report
  description: Operations with revenue report
paths:
  /accounts/{user_id}:
    get:
      tags:
      - account
      summary: Get user account
      operationId: getAccount
      parameters:
      - name: user_id
        in: path
        description: user id in uuid format
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountAggregate'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Account Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
      - account
      summary: Deposit funds to the user account
      operationId: depositFunds
      parameters:
      - name: user_id
        in: path
        description: user id in uuid format
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
      requestBody:
        $ref: '#/components/requestBodies/DepositFundsRequestBody'
      responses:
        "200":
          description: Funds added to the account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositFundsResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /accounts/{user_id}/transactions:
    post:
      tags:
      - account
      summary: Get list of account balance transactions
      operationId: getAccountTransactions
      parameters:
      - name: user_id
        in: path
        description: user id in uuid format
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
      requestBody:
        $ref: '#/components/requestBodies/TransactionListRequestBody'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionList'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /reserve/{user_id}:
    post:
      tags:
      - reserve
      summary: Reserve funds from user account
      operationId: reserveFunds
      parameters:
      - name: user_id
        in: path
        description: user id in uuid format
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
      requestBody:
        $ref: '#/components/requestBodies/ReserveFundsRequestBody'
      responses:
        "200":
          description: Funds reserved on the account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReserveFundsResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Account Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /reserve/{user_id}/revenue:
    post:
      tags:
      - reserve
      summary: Declare revenue
      description: Writes off money from the reserve wallet and adds it to the accounting report
      operationId: declareRevenue
      parameters:
      - name: user_id
        in: path
        description: user id in uuid format
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
      requestBody:
        $ref: '#/components/requestBodies/DeclareRevenueRequestBody'
      responses:
        "200":
          description: Added to revenue report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeclareRevenueResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Wallet Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /reports/{month}:
    get:
      tags:
      - report
      summary: Get revenue report
      description: test
      operationId: getRevenueReport
      parameters:
      - name: month
        in: path
        description: month of specific year in format "2022-12"
        required: true
        style: simple
        explode: false
        schema:
          pattern: ^\d{4}\-(0?[1-9]|1[012])$
          type: string
          example: 2022-10
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevenueReportResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      required:
      - details
      - message
      - status
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: string
    Account:
      required:
      - balance
      - id
      - user_id
      type: object
      properties:
        id:
          type: string
          format: uuid
          x-go-name: ID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        user_id:
          type: string
          format: uuid
          x-go-name: UserID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        balance:
          type: string
          example: "49.54"
    AccountAggregate:
      required:
      - balance
      - id
      - reserved
      - user_id
      type: object
      properties:
        id:
          type: string
          format: uuid
          x-go-name: ID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        user_id:
          type: string
          format: uuid
          x-go-name: UserID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        balance:
          type: string
          example: "49.54"
        reserved:
          type: string
          example: "33.31"
    TransactionOperation:
      type: string
      enum:
      - WITHDRAW
      - DEPOSIT
    Transaction:
      required:
      - amount
      - comment
      - from
      - id
      - operation
      - user_id
      type: object
      properties:
        id:
          type: string
          format: uuid
          x-go-name: ID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        user_id:
          type: string
          format: uuid
          x-go-name: UserID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        comment:
          type: string
          example: deposit from billing
        from:
          type: string
          description: empty means operation was initiated by billing
          format: uuid
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        amount:
          type: string
          example: "33.78"
        operation:
          $ref: '#/components/schemas/TransactionOperation'
    TransactionList:
      required:
      - next_page_cursor
      - transactions
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        next_page_cursor:
          type: string
          example: U3dhZ2dlciByb2Nrcw==
    DepositFundsRequest:
      required:
      - amount
      type: object
      properties:
        amount:
          type: string
          example: "35.47"
    ReserveFundsRequest:
      required:
      - amount
      - order_id
      - service_id
      type: object
      properties:
        service_id:
          type: string
          format: uuid
          x-go-name: ServiceID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        order_id:
          type: string
          format: uuid
          x-go-name: OrderID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        amount:
          type: string
          example: "75.48"
    DepositFundsResponse:
      required:
      - balance
      type: object
      properties:
        balance:
          type: string
          example: "49.54"
    ReserveFundsResponse:
      required:
      - account_balance
      - is_declared
      - order_id
      - reserved_amount
      - reserved_at
      - service_id
      type: object
      properties:
        reserved_amount:
          type: string
          example: "75.48"
        account_balance:
          type: string
          example: "75.48"
        is_declared:
          type: boolean
          x-go-name: Declared
        reserved_at:
          type: string
          format: datetime
          x-go-type: time.Time
        declared_at:
          type: string
          format: datetime
          x-go-type: time.Time
        service_id:
          type: string
          format: uuid
          x-go-name: ServiceID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        order_id:
          type: string
          format: uuid
          x-go-name: OrderID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
    DeclareRevenueRequest:
      required:
      - amount
      - order_id
      - service_id
      type: object
      properties:
        service_id:
          type: string
          format: uuid
          x-go-name: ServiceID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        order_id:
          type: string
          format: uuid
          x-go-name: OrderID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        amount:
          type: string
          example: "75.48"
    DeclareRevenueResponse:
      required:
      - account_id
      - declared_amount
      - is_declared
      - order_id
      - revenue_report_id
      - service_id
      type: object
      properties:
        service_id:
          type: string
          format: uuid
          x-go-name: ServiceID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        order_id:
          type: string
          format: uuid
          x-go-name: OrderID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        declared_amount:
          type: string
          example: "75.48"
        is_declared:
          type: boolean
          x-go-name: Declared
        declared_at:
          type: string
          format: datetime
          x-go-type: time.Time
        revenue_report_id:
          type: string
          format: uuid
          x-go-name: RevenueReportID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
    RevenueReportResponse:
      required:
      - report_url
      type: object
      properties:
        report_url:
          type: string
          format: uri
          x-go-name: ReportURL
    SortOrder:
      type: string
      enum:
      - ASC
      - DESC
    TransactionListSort:
      required:
      - amount
      - date
      type: object
      properties:
        date:
          $ref: '#/components/schemas/SortOrder'
        amount:
          $ref: '#/components/schemas/SortOrder'
    TransactionListRequest:
      required:
      - page_cursor
      - sort
      type: object
      properties:
        sort:
          $ref: '#/components/schemas/TransactionListSort'
        page_cursor:
          type: string
          example: U3dhZ2dlciByb2Nrcw==
  requestBodies:
    DepositFundsRequestBody:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DepositFundsRequest'
      required: true
    ReserveFundsRequestBody:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ReserveFundsRequest'
      required: true
    DeclareRevenueRequestBody:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DeclareRevenueRequest'
      required: true
    TransactionListRequestBody:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TransactionListRequest'
      required: true